pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Marcus Schiesser <mail@marcusschiesser.de>"]
readme = "README.md"
packages = [{include = "backend/app"}]

[tool.poetry.dependencies]
python = "^3.11,<3.12"
fastapi = "^0.104.1"
uvicorn = { extras = ["standard"], version = "^0.23.2" }
pypdf = "^3.17.0"
python-dotenv = "^1.0.0"
docx2txt = "^0.8"
llama-hub = "^0.0.77"
wikipedia = "^1.4.0"
llama-index-embeddings-huggingface = "^0.1.1"
llama-index-vector-stores-chroma = "^0.1.1"
llama-index-llms-ollama = "^0.1.1"
llama-index = "^0.10.4"
llama-index-vector-stores-postgres = "^0.1.1"
transformers = "^4.38.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
